Fixed movement bug
Because the direction of the collision was being determined by the width and height of the intersection, if you fell with enough speed at the corner of a tile you could clip into it. Since horizontal collisions aren't checked when velocity is 0 (because that breaks it), this let you stand partially inside a tile. Now you just stay above the tile like you should.